Move canonicalize_name to format
It's a formatting function after all...
